pyproject.toml
[tool.poetry]
name = "pytezos"
version = "3.0.4"
description = "Python toolkit for Tezos"
license = "MIT"
authors = ["Michael Zaikin <mz@baking-bad.org>", "Arthur Breitman", "Roman Serikov"]
readme = "README.md"
repository = "https://github.com/baking-bad/pytezos"
homepage = "https://pytezos.org"
keywords = ['tezos', 'blockchain', 'sdk', 'michelson', 'repl', 'cryptocurrencies', 'smart-contracts']
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [
    { include = "pytezos", from = "src" },
    { include = "pytezos_dapps", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
pyblake2 = "^1.1.2"
base58 = "^1.0.3"
requests = "^2.21.0"
pendulum = "*"
loguru = "*"
netstruct = "*"
tqdm = "*"
ply = "*"
simplejson = "*"
mnemonic = "*"
pyyaml = "*"
pysha3 = "1.0.2"
pysodium = "0.7.7"
secp256k1 = "0.13.2"
fastecdsa = "1.7.5"
bson = "^0.5.10"
strict_rfc3339 = "0.7"
py_ecc = "*"
deprecation = "*"
testcontainers = "^3.2.0"
click = "^7.1.2"
typing-extensions = "^3.7.4"
tortoise-orm = "^0.17.0"
aiohttp = "^3.7.4"
"ruamel.yaml" = "^0.17.0"
cattrs-extras = "^0.1.1"
websockets = "^8.1"
signalrcore = {git = "https://github.com/droserasprout/signalrcore", rev = "async"}
asyncpg = "^0.22.0"
pyee = "^8.1.0"

[tool.poetry.dev-dependencies]
parameterized = "^0.7.0"
pytest = "^3.0"
pytest-cov = "^2.4"
sphinx = "*"
sphinx_rtd_theme = { git = "https://github.com/readthedocs/sphinx_rtd_theme.git", tag = "0.5.0rc2" }
sphinx-sitemap = "*"
sphinxcontrib-googleanalytics = "^0.1"
bump2version = "^1.0.1"
pylint = "^2.7.2"
pylint-exit = "^1.2.0"
isort = "^5.7.0"
mypy = "^0.812"
diff-cover = "^4.2.3"

[tool.poetry.scripts]
pytezos = 'pytezos.cli.cli:cli'

[tool.isort]
line_length = 140

[build-system]
requires = ["poetry_core>=1.0.0", "cryptography==3.3.2", "wheel"]
build-backend = "poetry.core.masonry.api"
